back-end/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
databricks-sdk==0.38.0
msal==1.31.1
httpx==0.28.1
python-dotenv==1.0.1
openai>=1.0.0
